// types/blogTypes.ts
export type BlogItem = {
  title: string;
  description: string;
  id: number;
  image_name: string;
  audio_name: string;
  content: string;
  title_podcast: string;
  description_podcast: string;
  author_podcast: string;
  category_podcast: string;
  cover_podcast: string;
  create_at: string
  update_at: string
};